fix: allow alias declaration in typedstruct blocks
I had introduced this scope at some time, maybe due to some soundness bug in
Elixir that lead to leaks. Apparently removing it now does not break the test
ensuring the TypedStruct macros do not leak.

Due to this scope, any alias declared in the typedstruct block would not be
available during the type definition.

Fixes #22