bugfix: empty plugins now compile
Before, empty plugins injected an overrideable field/4 that would call
to field/3 if it existed. This worked fine if there actually was a
field/3. But if there wasn't, you'd get a compile error.

I've fixed that now with a before_compile callback. The code checks for the
existence of field/3 and /4, then inject the correct field/4.