test: fix tests on Elixir 1.14+
Since Elixir 1.14, it is not possible anymore to extract the type from a module
compiled in memory. This commit fixes this by defining a new
`TypedStruct.TestStruct` module in `test/support/test_struct.ex` containing all
test structs.

We take the opportunity to better name and organise the test structs in this
newly-created file.

This commit also fixes a bug in the `standardise/2` function, which clauses
where not properly ordered.

Refs: #31